Split reducers up
