+++ client/app/bundles/course/lesson_plan/reducers/milestones.jsx
@@ -1,0 +1,12 @@
+import Immutable from 'immutable';
+
+export const initialState = Immutable.fromJS([]);
+
+export default function (state = initialState, action) {
+  const { type } = action;
+
+  switch (type) {
+    default:
+      return state;
+  }
+}
